+++ src/modules/kiosks/components/DetailLoadCellsSide.js
@@ -1,15 +1,8 @@
 import React from 'react';
-import { Grid, Icon, Button, Segment, Image, Divider } from 'semantic-ui-react';
+import { Grid, Icon, Button, Segment } from 'semantic-ui-react';
 
 import { ReactComponent as NoImg } from 'styling/assets/images/noImg.svg';
 import ColoredBlock from 'modules/shared/components/ColoredBlock';
-
-const mystyle = {
-  paddingTop: '10px',
-  paddingLeft: '2px',
-  paddingRight: '2px',
-  minWidth: 80,
-};
 
 const Card = ({
   cellId,
@@ -17,8 +10,8 @@ const Card = ({
   productLine,
   totalProducts,
   handleEdit,
+  isActive,
 }) => {
-  const width = window.innerWidth;
   const handleClick = () => {
     handleEdit({
       productLine,
@@ -27,107 +20,48 @@ const Card = ({
       availableProducts: totalProducts,
     });
   };
-  if (width < 600) {
-    return (
-      <>
-        <Grid.Column style={mystyle}>
-          <Segment>
-            <div
-              style={{
-                marginLeft: '7px',
-                marginRight: '3px',
-              }}
-            >
-              <Grid.Row centered>
-                <button
-                  type="button"
-                  className="edit-icon"
-                  onClick={handleClick}
-                >
-                  <Icon name="edit" size="small" />
-                </button>
-                <b className="load-cell-position">
-                  {planogramPosition ? planogramPosition : 'Error'}
-                </b>
-                <p className="load-cell-calbeId">
-                  ID: {cellId ? cellId : 'Error'}
-                </p>
-              </Grid.Row>
-              <Grid.Row centered>
-                {productLine.image ? (
-                  <Image
-                    src={productLine.image}
-                    alt={productLine.name}
-                    size="mini"
-                  />
-                ) : (
-                  <NoImg />
-                )}
-              </Grid.Row>
 
-              <Grid.Row centered>
-                <b className="load-cell-title">
-                  {productLine ? `${productLine.name}` : null}
-                </b>
-              </Grid.Row>
-              <Grid.Row centered>
-                <p>
-                  Qty:&nbsp;
-                  <ColoredBlock type="b" value={totalProducts ? 100 : 0}>
-                    {totalProducts}
-                  </ColoredBlock>
-                </p>
-              </Grid.Row>
-              <Grid.Row centered>
-                <p>{`€ ${productLine.price}`}</p>
-              </Grid.Row>
-            </div>
-          </Segment>
-        </Grid.Column>
-      </>
-    );
-  } else {
+  if(isActive === false) {
     return (
-      <>
-        <Grid.Column className="load-cell">
-          <div className="load-cell-content">
-            <div className="load-cell-tags">
-              <button type="button" className="edit-icon" onClick={handleClick}>
-                <Icon name="edit" size="small" />
-              </button>
-              <b className="load-cell-position">
-                {planogramPosition ? planogramPosition : 'Error'}
-              </b>
-              <p className="load-cell-calbeId">
-                ID: {cellId ? cellId : 'Error'}
-              </p>
-            </div>
-            <div className="load-cell-info">
-              <div className="load-cell-icon">
-                {productLine.image ? (
-                  <img src={productLine.image} alt={productLine.name} />
-                ) : (
-                  <NoImg />
-                )}
-              </div>
-              <div className="load-cell-body">
-                <b className="load-cell-title">
-                  {productLine ? `${productLine.name}` : null}
-                </b>
-                <p>
-                  Qty:&nbsp;
-                  <ColoredBlock type="b" value={totalProducts ? 100 : 0}>
-                    {totalProducts}
-                  </ColoredBlock>
-                </p>
-                <p>{`€ ${productLine.price}`}</p>
-              </div>
-            </div>
-          </div>
-        </Grid.Column>
-      </>
+      <Grid.Column className="load-cell"></Grid.Column>
     );
   }
+
+  return (
+    <Grid.Column className="load-cell">
+      <div className="load-cell-content">
+        <div className="load-cell-tags">
+          <button type="button" className="edit-icon" onClick={handleClick}>
+            <Icon name="edit" size="small" />
+          </button>
+          <b className="load-cell-position">{planogramPosition ? planogramPosition : "Error"}</b>
+          <p className="load-cell-calbeId">ID: {cellId ? cellId : "Error"}</p>
+        </div>
+        <div className="load-cell-info">
+          <div className="load-cell-icon">
+            {productLine.image ? (
+              <img src={productLine.image} alt={productLine.name} />
+            ) : (
+                <NoImg />
+              )}
+          </div>
+          <div className="load-cell-body">
+            <b className="load-cell-title">
+              {productLine ? `${productLine.name}` : null}
+            </b>
+            <p>
+              Qty:&nbsp;
+              <ColoredBlock type="b" value={totalProducts ? 100 : 0}>
+                {totalProducts}
+              </ColoredBlock>
+            </p>
+            <p>{`€ ${productLine.price}`}</p>
+          </div>
+        </div>
+
+      </div>
+    </Grid.Column>
+  );
 };
 
 const DetailLoadCellsSide = ({ cells, handleEdit, handleAdd }) => (
@@ -143,15 +77,9 @@ const DetailLoadCellsSide = ({ cells, handleEdit, handleAdd }) => (
                 //   fullWidth = Boolean(props);
                 // }
                 // if ((fullWidth && j) || (!fullWidth && !j)) return null;
-                return (
-                  props && (
-                    <Card
-                      {...props}
-                      key={props.cellId}
-                      handleEdit={handleEdit}
-                    />
-                  )
-                );
+                return props && (
+                  <Card {...props} key={props.cellId} handleEdit={handleEdit} />
+                )
                 //  (
                 //     <Grid.Column
                 //       key={performance.now().toString(32)}