+++ src/modules/kiosks/components/KiosksContent.js
@@ -36,8 +36,7 @@ const columns = [
     formatter: ({ serialNumber }) => {
       if (serialNumber.length > 20) {
         return serialNumber.substring(0, 15) + '...';
-      }
-      else return serialNumber
+      } else return serialNumber;
     },
   },
   {
@@ -64,11 +63,7 @@ const columns = [
     field: 'location',
     formatter: ({ location: { address } }) => {
       const { postalCode, city } = address;
-      const addr = [
-        postalCode,
-        city,
-        !postalCode && !city && 'N.A.'
-      ]
+      const addr = [postalCode, city, !postalCode && !city && 'N.A.']
         .filter(el => Boolean(el))
         .join(', ');
       return addr;
@@ -77,7 +72,13 @@ const columns = [
   {
     title: 'Sales Today',
     field: 'dayIncome',
-    formatter: ({ dayIncome }) => `€ ${dayIncome}`,
+    // formatter: ({ dayIncome }) => `€ ${dayIncome}`,
+    formatter: ({ dayIncome }) => {
+      if (dayIncome === '') {
+        return '';
+      }
+      return <div style={{ textAlign: 'right' }}> {dayIncome}€ </div>;
+    },
   },
 ];
 
@@ -123,7 +124,6 @@ const KiosksContent = ({
         ...kio,
         ...door,
       });
-      data.skip = 0;
     }
 
     if (sort && sortValue[sort[0].column]) {