+++ CHANGELOG.md
@@ -1,3 +1,117 @@
+#### 1.3.23 (2021-03-03)
+
+##### Build System / Dependencies
+
+* **LIV-2833:**  remove the mandatory for address field (00a4539f)
+* **LIV-2794:**  resolved pagination and added loading screen (efa026cd)
+
+##### New Features
+
+* **Edit/Delete Loadcell:** Merge into LIV-2327 (b47760a7)
+* **delete loadcell:** LIV-1800 delete load cell (535728de)
+* **Kiosk-LoadCell Modal:**  LIV-2327 Cable ID should be editable (c1cd3194)
+
+##### Bug Fixes
+
+* **LIV-2794:**  Datepicker format is changed (5b5b636f)
+* **Date Picker:**  LIV-2625: Datepicker length updated (f98b24a9)
+* **LIV-2833:**  Make the address field required in user form (3ad06d59)
+* **Kiosk:**  LIV-2647: Kiosk Info Address formate Updated (d4e009d9)
+* **Dashboard:**  LIV-2723: log statements removed (c3ed7e6e)
+* **kiosk details Mobile Responsive:**  LIV-2723  Koisk Detail and sub components (3d72aadf)
+
+#### 1.3.22 (2021-02-26)
+
+##### Build System / Dependencies
+
+* **LIV-2805:**  hidden the add kiosk btn for admin and validated serialnumber on creation (8390b2c9)
+
+##### New Features
+
+* **datepicker:** LIV-2725 updated date picker component for mobile view (d42dfd5a)
+
+##### Bug Fixes
+
+* **Dashboard:**  LIV-2725: Margin issue resolved (9a3d8161)
+* **deshboard Kiosk Transactions:**  LIV-2725:Searchbox Dropdown and Button Responsiveness feature is added (b66dd3e4)
+* **deshboard Kiosk Products Transactions Refils and Sales:**  LIV-2793 Formate of Table Updated (d5b833b5)
+
+##### Other Changes
+
+* livello-network/livello-mission-control into staging-master (9fbf6f2a)
+
+#### 1.3.21 (2021-02-23)
+
+##### Bug Fixes
+
+* **transactions:**  swap column order for 'Article Number' (daf82245)
+* **Transactions:**
+  *  LIV-2827: Article Number Column is added in Sales and Refills Table (2a1c27b4)
+  *  LIV-2827: Article Number Column is added in Sales and Refills Table (29c030f0)
+
+#### 1.3.20 (2021-02-18)
+
+##### Build System / Dependencies
+
+* **LIV-2795:**  added 3 more columns to alerts table (f1527f90)
+* **liv-2591:**  change landing page (1653ab84)
+
+##### New Features
+
+* **refills:**  update grid refills cost (8c9251e8)
+
+##### Bug Fixes
+
+* **Dashboard Alert:**
+  *  LIV-2824: Hide Duration, Severity and Status Column in Show all Table (ea53ee2c)
+  *  LIV-2824: Hide Duration, Severity and Status Column (e844d942)
+* **Kiosk Settings Tooltip:**  LIV-2790: Tooltip added for default support language (f9ede372)
+* **LIV-2591:**  logging out renewing the token instantly resolved (a923376d)
+* **kiosk Settings:**  LIV-2672 Model Msg Updated (87b95fd2)
+* **LIV-2751:**  PackageOptions fetch old data bug solved (53ea3181)
+
+#### 1.3.19 (2021-02-15)
+
+##### Build System / Dependencies
+
+* **LIV-2795:**  added 3 more columns to alerts table (f1527f90)
+* **liv-2591:**  change landing page (1653ab84)
+
+##### New Features
+
+* **refills:**  update grid refills cost (9550350b)
+
+##### Bug Fixes
+
+* **refills:**  cost - number of signs after comma fixed to 2 (7071cf9b)
+* **LIV-2591:**  logging out renewing the token instantly resolved (a923376d)
+* **kiosk Settings:**  LIV-2672 Model Msg Updated (87b95fd2)
+* **LIV-2751:**  PackageOptions fetch old data bug solved (53ea3181)
+
+#### 1.3.18 (2021-02-01)
+
+##### Bug Fixes
+
+* **kiosk settings:**  LIV-2681 member card tool tip updated (48bb727d)
+
+#### 1.3.17 (2021-02-01)
+
+##### Bug Fixes
+
+* **kioskProps:** LIV-2681 planogram fixed after updating kiosk settings (20635245)
+* **kiosk settings:**  age restirction warning message chaned to german (67004d0f)
+
+#### 1.3.16 (2021-01-29)
+
+##### Bug Fixes
+
+* **kiosk settings:**  warning color, out of service modal no confirmation (49296f3c)
+* **liv-2717:**  membercard issue to revert back when there's no change (2d967165)
+
+##### Other Changes
+
+* **dashboard,kiosk,transactions:**  update the name of kiosks. (d46ee5b4)
+
 #### 1.3.15 (2021-01-26)
 
 ##### Bug Fixes